Optimization: reassign minimized values before updating
I completely missed this. This means that each update was using the
previous iteration's values so this should actually improve results.

Also, a superfluous variable probably left over from one of the early
design iterations has been deleted and the Lagrangian now has an
exclamation mark to denote that it actually mutates the layout since it
calls the update! function. What is shadowed has also been adjusted for
consistency and some names changed for clarity.