Vector refraction: build coordinate vectors through the trace
This seems like it will be necessary to complete the pupil sampling in
an accurate way. The vector refraction method does not return the full
trace data including the angles in order to allocate less & remain a bit
more efficient as it's called in a hot loop, which is why it was only
previously returning the end points. However, from the looks of it I
think I'm going to have to switch the full_trace method to sample
Cartesian nodes and use rejecting sampling in order to capture the full
pupil due to the non-linearity of the mapping. For this the values at
the aperture stop need to be checked and so the vector refraction method
needs to contain that data.