Pupil sampling: accurately bound the trace by finding the edge rays
This should improve results by covering the full pupil. Note that
automatic differentiation cannot be used as both the RealRay trace
function and the vectorial method do not support dual numbers due to
type conversions so the computed derivatives are not exact as Optim will
default to using centralized difference quotients for the BFGS
algorithm; that should be fine as the initial guesses should be close.

This still might not be entirely accurately as I have the sense that it
could be possible for outer rays past this bound to bend back into the
pupil and maybe for some unusual systems the bracket won't be large
enough; I'm not sure as the non-linearity of the system makes it
difficult to reason about, maybe there's some theorem somewhere.