Tests: leave out pupil sampling tests which are no longer valid
Since the move towards Cartesian rejection sampling these can no longer
be tested because the coordinates won't match. The way to test this now
would be to argmax the radial coordinate from the ray error struct and
try to find the pi / 2 angular coordinate, but this will not be exact
and it's more trouble than it's worth. The pupil sampling tests are
visual for the most part -- I've been testing with spot diagrams and
fitting the geometric data to Zernike polynomials and plotting the
wavefront error. Without bringing my base Zernike package into the tests
and checking coefficients there's not a great way to test this other
than visual inspection. A good way, however, is to use a known result
and compare the RMS spot sizes, which is what I'll do soon.