Surface plot: guard negative radicands on z-coordinate computation
For a specified aperture size larger than the radius of curvature the
maximum axial coordinate should max out at the radius. These apertures
are of course not physical, but it's better than running into a generic
error, especially since it's through the Makie extension and the error
plus the stacktrace are not descriptive. At the moment the surfaces are
drawn using spherical geometry, even though the package now supports
aspheric surfaces, so the base radius at the vertex could differ I
suppose and cause issues; I suppose it wouldn't be all too difficult to
at least bring the conic constant in and draw them accurately, but for
now this is fine. I might implement that later after I finish the full
trace feature, but the problem with that is the support for the surface
polynomial functions won't be taken into account; I still think it's
worth it to at least have accurate visual surface traces for parabolic
and hyperbolic surfaces for telescopic systems etc.; for that the
quotient form of the sagittal equation would be used and the
y-coordinate would be calculated by compensating z by 1 + K and dividing
the radicand by 1 + K.