Open pupil sampling for input `RayBasis`, implement through-focus
The full_trace function under RayBasis has been re-implemented /
properly implemented and the focus position can now be specified.

This should complete the pupil sampling feature assuming everything is
more or less correct.

I'm going to rework the System interface across the entire package in
order to simplify everything so some of the dispatching introduced here
is going to change.