Pupil sampling: test surface extension, test a singlet
I found a book lying around which had this example along with the spot
diagrams and RMS data for the three different fields. The thing with the
pupil sampling tests is that it's also visual with regards to the spot
diagrams, and that the RMS values could differ depending on how many
rays are traced and the pupil sampling technique used. Like I've
mentioned before ideally our pupil sampling functionality should be
compared with results from existing optical design software, but I
currently don't have access to any of those so this will do for now.