Surface plot: support drawing aspheric surfaces
This takes the conic constant into account.

Negative radicands for the y-coordinate are not guarded as I'm pretty
sure the radicand should never be negative here given how z_max is
calculated although I've not done any strict analysis on whether there
would be some edge cases; in any case, it's preferable to have an
exception thrown here rather than silently setting it to a NaN value or
taking an absolute value and drawing the surface incorrectly. Also, it
allows for a more efficient computation by leveraging broadcast fusion
without the need for intermediary allocations.