Rework the `System` interface, include the `Layout` in the struct
This should simplify things and allow many of the functions to simply
accept a System rather than a surfaces + system combo. The fat struct
design for this package makes the interface very simple: you construct
an optical system either with a matrix or the Layout constructor,
specify the aperture sizes and image height which determine the lens
speed and field of view, and then use the result from the solve function
to analyze everything else.

Note that the added complexity comes from the fact that the solve
function accepts a directly constructed Lens and so the surface layout /
prescription will be empty in that case so dispatch is used here to
instantiate the necessary system depending on the layout.

This includes a fix for the real chief ray function which wasn't taking
into account aspheric parameters. Maybe that should be its own commit,
but it makes sense to include it here as it fits the theme.